requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1
marshmallow==3.20.1
pymysql==1.1.0
marshmallow-sqlalchemy==0.29.0
flask-sqlalchemy==3.1.1
email-validator==2.0.0
flask-migrate==4.0.5
python-dateutil==2.8.2 